backend-python/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ollama
python-dotenv